filterting + delete

+++ js/apps/mail/service/email-service.js
@@ -12,7 +12,8 @@ export const emailService = {
     query,
     getById,
     removeEmail,
-    updateEmail
+    updateEmail,
+    sendEmail
 
 }
 function updateEmail(email) {
@@ -21,7 +22,18 @@ function updateEmail(email) {
 
 
 function removeEmail(emailId) {
-    return storageService.remove(EMAILS_KEY, emailId)
+    return getById(emailId)
+        .then(email => {
+            if (email.isDeleted) {
+                console.log('deleted', email)
+                return storageService.remove(EMAILS_KEY, emailId);
+            }
+            else {
+                email.isDeleted = true;
+                console.log('trashed', email)
+                return storageService.put(EMAILS_KEY, email)
+            }
+        })
 }
 
 
@@ -34,6 +46,19 @@ function query() {
 
 function getById(emailId) {
     return storageService.get(EMAILS_KEY, emailId);
+}
+
+function sendEmail(email) {
+    email.id = utilService.makeId(4);
+    email.from = loggedUser.fullName;
+    email.fromEmail = loggedUser.email;
+    email.isRead = false;
+    email.isSent = true;
+    email.isStarred = false;
+    email.isDeleted = false;
+    email.sentAt = Date.now();
+    return storageService.post(EMAILS_KEY, email);
+
 }
 
 function _createEmails() {
